+++ Sources/MapboxMaps/Annotations/AnnotationImagesManager.swift
@@ -15,10 +15,16 @@ extension PointAnnotation {
     public struct Image: Hashable, Sendable {
         public var image: UIImage
         public var name: String
+        public var sdf: Bool
 
         public init(image: UIImage, name: String) {
+            self.init(image: image, name: name, sdf: false)
+        }
+
+        public init(image: UIImage, name: String, sdf: Bool) {
             self.image = image
             self.name = name
+            self.sdf = sdf
         }
     }
 }